+++ bidict/metadata.py
@@ -7,7 +7,7 @@
 """Define bidict package metadata."""
 
 
-__version__ = '0.22.1.dev0'
+__version__ = '0.22.1'
 __author__ = {'name': 'Joshua Bronson', 'email': 'jabronson@gmail.com'}
 __copyright__ = '© 2009-2022 Joshua Bronson'
 __description__ = 'The bidirectional mapping library for Python.'